requirements.txt
aiohttp==3.9.3
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.7.2
attrs==23.2.0
blinker==1.7.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.2
defusedxml==0.7.1
Django==5.0.2
django-allauth==0.61.1
Flask==3.0.2
frozenlist==1.4.1
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
multidict==6.0.5
oauthlib==3.2.2
pycparser==2.21
PyJWT==2.8.0
pyngrok==7.1.2
python-dotenv==1.0.1
python3-openid==3.2.0
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
sqlparse==0.4.4
twilio==8.13.0
tzdata==2024.1
urllib3==2.2.0
Werkzeug==3.0.1
yarl==1.9.4
